Move microphone visualizer out of user interface namespace because this componnet only used in the setting screen.
